+++ src/components/PostWeather.js
@@ -1,7 +1,9 @@
 import moment from "moment";
 import { Link } from "react-router-dom";
+import { useFetch } from "../Utills/useFetch";
 
-const PostWeather = ({ arr, conversion, getDayDetails }) => {
+const PostWeather = () => {
+  const { getDayDetails, conToCelcius, arr } = useFetch();
   const newArr = arr.filter(
     (datas) => moment.unix(datas.dt).format("LL") != moment().format("LL")
   );
@@ -17,7 +19,7 @@ const PostWeather = ({ arr, conversion, getDayDetails }) => {
                 <div className="row">
                   <div className="days" onClick={() => getDayDetails(index)}>
                     <h3>Date: {moment.unix(data.dt).format("LL")}</h3>
-                    <h3>Temp: {conversion(data.temp.day)}&deg;</h3>
+                    <h3>Temp: {conToCelcius(data.temp.day)}&deg;</h3>
                     <h2 className="description">
                       Description: {data.weather[0].description}
                     </h2>
