+++ src/screens/DetailsScreen.js
@@ -1,0 +1,10 @@
+import Details from "../components/Details";
+import { useFetch } from "../Utills/useFetch";
+import { minmaxTemp } from "../Utills/Functions";
+import { useEffect } from "react";
+
+const DetailsScreen = () => {
+  return <Details />;
+};
+
+export default DetailsScreen;
